Restructure project, update README, change unit tests accordingly
